Rename files of table form
